backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
bcrypt==4.0.1
marshmallow==3.20.1
python-dotenv==1.0.0
Pillow==10.4.0
scikit-learn==1.5.2
numpy==1.24.3
pandas==2.0.3